+++ src/category/mod.rs
@@ -1,3 +1,0 @@
-#![allow(deprecated)]
-
-pub mod function_category;